타이머 테스트용 출력 삭제
